requirements.txt
requests
watchdog
python-dotenv
numpy
Pillow